src/com/amazon/ata/cost/MonetaryCostStrategy.java
package com.amazon.ata.cost;

import com.amazon.ata.types.Material;
import com.amazon.ata.types.Packaging;
import com.amazon.ata.types.ShipmentOption;

import java.math.BigDecimal;
import java.util.HashMap;
import java.util.Map;

/**
 * Considers both material cost and labor cost to calculate the total cost of a ShipmentOption.
 */
public class MonetaryCostStrategy implements CostStrategy {

    private final Map<Material, BigDecimal> materialCostPerGram;
    private static final BigDecimal LABOR_COST = BigDecimal.valueOf(0.43);

    /**
     * Initializes a MonetaryCostStrategy.
     */
    public MonetaryCostStrategy() {
        materialCostPerGram = new HashMap<>();
        materialCostPerGram.put(Material.CORRUGATE, BigDecimal.valueOf(.005));
        materialCostPerGram.put(Material.LAMINATED_PLASTIC, BigDecimal.valueOf(.25));
    }

    @Override
    public Currency getCost(ShipmentOption shipmentOption) {
        Packaging packaging = shipmentOption.getPackaging();
        BigDecimal materialCost = this.materialCostPerGram.get(packaging.getMaterial());

        BigDecimal cost = packaging.getMass().multiply(materialCost).add(LABOR_COST);
        return new Currency(cost);
    }
}